make it be able to build on macOS

+++ src/game/server/entities/projectile.cpp
@@ -113,8 +113,8 @@ void CProjectile::Tick()
 				int y = GameServer()->Collision()->getTeleY(0);
 				int tx = GameServer()->Collision()->getTeleX(1);
 				int ty = GameServer()->Collision()->getTeleY(1);
-				vec2 start = {x, y};
-				vec2 end = {tx, ty};
+				vec2 start = {(float)x, (float)y};
+				vec2 end = {(float)tx, (float)ty};
 				m_Pos = m_Pos - start * 32 + end * 32;
 			}
 		} else if (GameServer()->Collision()->GetCollisionAt(CurPos.x, CurPos.y)&CCollision::COLFLAG_TELETWO) {
@@ -124,8 +124,8 @@ void CProjectile::Tick()
 				int y = GameServer()->Collision()->getTeleY(1);
 				int tx = GameServer()->Collision()->getTeleX(0);
 				int ty = GameServer()->Collision()->getTeleY(0);
-				vec2 start = {x, y};
-				vec2 end = {tx, ty};
+				vec2 start = {(float)x, (float)y};
+				vec2 end = {(float)tx, (float)ty};
 				m_Pos = m_Pos - start * 32 + end * 32;
 			}
 		} else if (GameServer()->Collision()->GetCollisionAt(CurPos.x, CurPos.y)&CCollision::COLFLAG_TELETHREE) {
@@ -135,8 +135,8 @@ void CProjectile::Tick()
 				int y = GameServer()->Collision()->getTeleY(2);
 				int tx = GameServer()->Collision()->getTeleX(3);
 				int ty = GameServer()->Collision()->getTeleY(3);
-				vec2 start = {x, y};
-				vec2 end = {tx, ty};
+				vec2 start = {(float)x, (float)y};
+				vec2 end = {(float)tx, (float)ty};
 				m_Pos = m_Pos - start * 32 + end * 32;
 			}
 		} else if (GameServer()->Collision()->GetCollisionAt(CurPos.x, CurPos.y)&CCollision::COLFLAG_TELEFOUR) {
@@ -146,8 +146,8 @@ void CProjectile::Tick()
 				int y = GameServer()->Collision()->getTeleY(3);
 				int tx = GameServer()->Collision()->getTeleX(2);
 				int ty = GameServer()->Collision()->getTeleY(2);
-				vec2 start = {x, y};
-				vec2 end = {tx, ty};
+				vec2 start = {(float)x, (float)y};
+				vec2 end = {(float)tx, (float)ty};
 				m_Pos = m_Pos - start * 32 + end * 32;
 			}
 		} else {
